# docs/api/openapi-spec.yaml
openapi: 3.1.0
info:
  title: AssuranceNet Document Management API
  description: |
    Azure-native document management API replacing Oracle Universal Content Manager (UCM)
    for FSIS AssuranceNet. Provides document upload, download, PDF conversion tracking,
    PDF merging, investigation management, and audit logging.

    ## FSIS Data Context
    This API manages food safety investigation documents from the
    [FSIS Science & Data program](https://www.fsis.usda.gov/science-data), including
    sampling plans, laboratory data (CSV), National Residue Program reports,
    microbiology baselines, MPI establishment directories, and inspection task records.

    ## Document Versioning
    Documents support versioning for compliance and audit requirements. Each document
    has a stable logical identity (document ID) and one or more versions (physical binaries).
    End users see only the latest version; administrators can access historical versions
    and rollback if needed.

    ## PDF Conversion
    The system uses Aspose.Words as the default PDF engine for document conversions.
    PDF conversion is tracked asynchronously; check `pdf_conversion_status` for
    progress (pending, processing, completed, failed, not_required).

    ## Merge Ordering
    PDF merges follow type-based ordering (FSIS-specific rules):
    1. Investigation Reports (10)
    2. Inspection Forms (20)
    3. Laboratory Results (30)
    4. Legal Documents (40)
    5. Correspondence (50)
...

...

...
  contact:
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Microsoft Entra ID JWT access token
  schemas:
...
    Investigation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        record_id:
          type: string
          example: "INVESTIGATION-12345"
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, closed, archived]
        created_by:
          type: string
        created_by_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        document_count:
          type: integer

    # ========================================================================
    # Document Version Schema (Physical Binary Metadata)
    # ========================================================================
...
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
...
paths:
...
  /api/v1/investigations:
    get:
      summary: List investigations
      tags: [Investigations]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, closed, archived]
      responses:
        "200":
          description: Paginated investigation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
    post:
      summary: Create investigation
      tags: [Investigations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [record_id, title]
              properties:
                record_id:
                  type: string
                title:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Investigation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
  /api/v1/investigations/{investigation_id}:
    get:
      summary: Get investigation details
      tags: [Investigations]
      parameters:
        - name: investigation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Investigation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
        "404":
          description: Investigation not found
  /api/v1/investigations/{investigation_id}/documents:
    get:
      summary: List documents for investigation
      tags: [Documents]
      parameters:
        - name: investigation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Document list (latest versions only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"

  # ========================================================================
  # Document Upload & Management Endpoints
  # ========================================================================
...
  /api/v1/investigations/{investigation_id}/merge-pdf:
    post:
      summary: Merge multiple document PDFs
      tags: [PDF Operations]
      description: |
        Merges the latest versions of multiple documents into a single PDF.
        Documents are ordered by FSIS type (investigation_report → inspection_form
        → laboratory_result → legal_document → correspondence → supporting_evidence
        → other), then by creation timestamp within each type.

        Request body uses document_ids (not file_ids).
      parameters:
        - name: investigation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_ids]
              properties:
                document_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 2
                  maxItems: 50
                  description: Document IDs to merge (will use latest version of each)
      responses:
        "200":
          description: Merged PDF download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid request (empty list, invalid IDs, documents not found, etc.)
        "422":
          description: Merge failed (PDF conversion pending, missing files, etc.)

  # ========================================================================
  # Admin Document Endpoints
  # ========================================================================
...